Правки 7 задание
