+++ css/style.css
@@ -69,6 +69,7 @@ fieldset {
 
 .map {
   position: relative;
+  z-index: 0;
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
