+++ js/util.js
@@ -18,4 +18,26 @@ let getRandomFixed = (min, max, decimal) => {
   }
 };
 
-export {getRandomArbitrary, getRandomFixed};
+const showError = (error) => {
+
+  const alertContainer = document.createElement('div');
+  alertContainer.style.zIndex = 1000;
+  alertContainer.style.position = 'fixed';
+  alertContainer.style.left = 0;
+  alertContainer.style.top = 0;
+  alertContainer.style.right = 0;
+  alertContainer.style.padding = '10px 3px';
+  alertContainer.style.fontSize = '30px';
+  alertContainer.style.textAlign = 'center';
+  alertContainer.style.backgroundColor = 'red';
+
+  alertContainer.textContent = error;
+
+  document.body.append(alertContainer);
+
+  setTimeout(() => {
+    alertContainer.remove();
+  }, 5000);
+}
+
+export {getRandomArbitrary, getRandomFixed, showError};
